refactor: 冗長になっていたroom_channelを分割して整理 #20
